Breaking Change(npm): rename to textlint-filter-rule-node-types (Fix #5)
