+++ src/textlint-filter-rule-node-types.js
