+++ test/textlint-filter-rule-node-types-test.js
@@ -2,12 +2,12 @@
 "use strict";
 const TextLintCore = require("textlint").TextLintCore;
 const TextLintNodeType = require("textlint").TextLintNodeType;
-import ignoreRule from "../src/textlint-rule-ignore-node-types";
+import filterRule from "../src/textlint-filter-rule-node-types";
 import reportRule from "textlint-rule-report-node-types";
 const assert = require("power-assert");
-describe("textlint-rule-ignore-node-types", function () {
-    context("when report type", function () {
-        it("should messages includes the type message", function () {
+describe("textlint-filter-rule-node-types", function () {
+    context("when report and ignore type", function () {
+        it("should messages is empty", function () {
             const textlint = new TextLintCore();
             textlint.setupRules({
                 report: reportRule
@@ -16,23 +16,10 @@ describe("textlint-rule-ignore-node-types", function () {
                     nodeTypes: [TextLintNodeType.Str]
                 }
             });
-            return textlint.lintText("text").then(({messages}) => {
-                assert.equal(messages.length, 1);
-                assert.equal(messages[0].message, TextLintNodeType.Str);
-            });
-        });
-    });
-    context("when report and ignore type", function () {
-        it("should messages is empty", function () {
-            const textlint = new TextLintCore();
-            textlint.setupRules({
-                ignore: ignoreRule,
-                report: reportRule
+            textlint.setupFilterRules({
+                ignore: filterRule
             }, {
                 ignore: {
-                    nodeTypes: [TextLintNodeType.Str]
-                },
-                report: {
                     nodeTypes: [TextLintNodeType.Str]
                 }
             });
@@ -41,21 +28,46 @@ describe("textlint-rule-ignore-node-types", function () {
             });
         });
     });
+    context("when fixer report and ignore type", function () {
+        it("should messages is empty", function () {
+            const textlint = new TextLintCore();
+            textlint.setupRules({
+                report: reportRule
+            }, {
+                report: {
+                    nodeTypes: [TextLintNodeType.Str]
+                }
+            });
+            textlint.setupFilterRubles({
+                ignore: filterRule
+            }, {
+                ignore: {
+                    nodeTypes: [TextLintNodeType.Str]
+                }
+            });
+            return textlint.fixText("text").then(({messages}) => {
+                assert.equal(messages.length, 0);
+            });
+        });
+    });
     context("when report multiple nodes and ignore type", function () {
         it("should messages is empty", function () {
             const textlint = new TextLintCore();
             textlint.setupRules({
-                ignore: ignoreRule,
                 report1: reportRule,
                 report2: reportRule
             }, {
-                ignore: {
-                    nodeTypes: [TextLintNodeType.Str]
-                },
                 report1: {
                     nodeTypes: [TextLintNodeType.Str]
                 },
                 report2: {
+                    nodeTypes: [TextLintNodeType.Str]
+                }
+            });
+            textlint.setupFilterRules({
+                ignore: filterRule
+            }, {
+                ignore: {
                     nodeTypes: [TextLintNodeType.Str]
                 }
             });
